split statement and bankaccount classes into separate files
